[Chore] Remove shared utility facades
