--- src/core/cleanup.rs
@@ -1,1 +1,0 @@
-pub use sekai_asset_pipeline::{remove_file_if_exists, remove_file_with_retries};
