--- src/core/codec.rs
@@ -1,6 +1,0 @@
-pub use sekai_asset_pipeline::{
-    codec_summary, decode_hca_bytes_to_wav, decode_hca_bytes_to_wav_bytes, decode_hca_to_wav,
-    export_acb, export_acb_unique_to_memory, export_usm, export_usm_reader_to_memory,
-    export_usm_to_memory, file_has_usm_magic, has_usm_magic, read_usm_metadata, CodecSummary,
-    CODEC_BACKEND,
-};
